Re-house lib classes under Gitlab::Composer
